Follow franciscomendes10866 tutorial
Followed franciscomendes10866's tutorial on Next.js, Tailwind,
TRPC, and Prisma.

Had to make a good bit of changes to make work with the newest
versions of Next.js, TRPC, and Prisma.

Main issue was that many instances of Prisma were being created the
way it was done in the tutorial until I pullsed the Prisma
instantiation into it's own file.

Tutorial here:
https://dev.to/franciscomendes10866/build-a-full-stack-app-with-nextjs-tailwind-trpc-and-prisma-orm-4ail